LayoutTests/fast/dom/Element/attr-param-typechecking-expected.txt

This test checks whether passing wrong types to setAttributeNode causes a crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.setAttributeNode(); threw exception TypeError: Not enough arguments.
PASS element.setAttributeNodeNS(); threw exception TypeError: Not enough arguments.
PASS element.removeAttributeNode(); threw exception TypeError: Not enough arguments.
PASS element.setAttributeNode("style"); threw exception TypeError: Type error.
PASS element.setAttributeNode(null); threw exception TypeError: Type error.
PASS element.setAttributeNode(undefined); threw exception TypeError: Type error.
PASS element.setAttributeNode(new Object); threw exception TypeError: Type error.
PASS element.removeAttributeNode("style"); threw exception TypeError: Type error.
PASS element.removeAttributeNode(null); threw exception TypeError: Type error.
PASS element.removeAttributeNode(undefined); threw exception TypeError: Type error.
PASS element.removeAttributeNode(new Object); threw exception TypeError: Type error.
PASS element.setAttributeNodeNS("style"); threw exception TypeError: Type error.
PASS element.setAttributeNodeNS(null); threw exception TypeError: Type error.
PASS element.setAttributeNodeNS(undefined); threw exception TypeError: Type error.
PASS element.setAttributeNodeNS(new Object); threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
